requirements.txt
gitpython==3.1.43
python-gitlab==3.12.0
argparse ==1.4.0